+++ src/services/api.js
@@ -188,56 +188,9 @@ export async function fetchPopularHashItems() {
     },
   ];
 }
-export const createPost = async (body) => {
-  await console.log(body);
-};
 
 export const createcomment = async (body) => {
   await console.log(body);
-};
-
-export const fetchPostdetail = async (postId) => {
-  return {
-    code: 1201,
-    message: "게시글 조회에 성공하였습니다.",
-    result: {
-      categoryInfo: {
-        id: 1,
-        title: "프로젝트 자랑 게시판",
-      },
-      memberInfo: {
-        Id: 0,
-        nickname: "MogensEgeskov",
-        development: "iOS Developer",
-        imageUrl: "~~~",
-      },
-      postInfo: {
-        title: "UI Templates",
-        content:
-          "<p>My first iOS app is available on the AppStore. I literally didn’t know anything about SwiftUI (still not much) and in probably 4 weeks was able to recreate my android app for iOS. Highly recommend MengTo video. My first iOS app is available on the AppStore. I literally didn’t know anything about SwiftUI (still not much) and in probably 4 weeks was able to recreate my android app for iOS. Highly recommend MengTo video. My first iOS app is available on the AppStore. I literally didn’t know anything about SwiftUI (still not much) and in probably 4 weeks was able to recreate my android app for iOS. Highly recommend MengTo video.My first iOS app is available on the AppStore. I literally didn’t know anything about SwiftUI (still not much) and in probably 4 weeks was able to recreate my android app for iOS. Highly recommend MengTo video. My first iOS app is available on the AppStore. I literally didn’t know anything about SwiftUI (still not much) and in probably 4 weeks was able to recreate my android app for iOS. </p>",
-        viewCount: 12,
-        commentCount: 12,
-        recommedCount: 12,
-        createdAt: new Date().toLocaleDateString(),
-        isModified: true,
-        hashtags: ["#test"],
-        imgUrl: [
-          {
-            id: 1,
-            url: "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcS6KVvlziiJYFxZZIq3Xc_dVuzIbSLrgvtHPA&s",
-          },
-          {
-            id: 2,
-            url: "https://www.shutterstock.com/ko/blog/wp-content/uploads/sites/17/2021/01/2021-graphic-design-banner.jpg",
-          },
-          {
-            id: 3,
-            url: "https://img1.daumcdn.net/thumb/R1280x0/?fname=http://t1.daumcdn.net/brunch/service/user/6H5a/image/VbC1Pts-64VW9-xWDV3qad5cLok.jpg",
-          },
-        ],
-      },
-    },
-  };
 };
 
 export const fetchcomment = async (postId) => {
